[refactor]: share > ui > HighlightText 컴포넌트 추가
